requirements.txt
opencv-python-headless
timm
pillow
grpcio-tools
grpcio
python-grpc
ijson
requests

transformers
torch==2.0.1a0
torchvision==0.15.2a0
intel_extension_for_pytorch==2.0.110+xpu
ultralytics
